model/requirements.txt
pandas
numpy
scikit-learn
matplotlib
jupyter
alpaca-trade-api
azureml-sdk
